Moved new testss to test_core.py and removed test_doppler_axis.py
